split out ktx and ktx 2 to seperate files
