+++ src/compressed-textures/ktx2/const.ts
