+++ src/compressed-textures/ktx2/init.ts
@@ -1,0 +1,9 @@
+import { extensions } from '../../extensions/Extensions';
+import { detectCompressed } from '../shared/detectCompressed';
+import { resolveCompressedTextureUrl } from '../shared/resolveCompressedTextureUrl';
+import { loadKTX2 } from './loadKTX2';
+
+extensions.add(loadKTX2);
+extensions.add(resolveCompressedTextureUrl);
+extensions.add(detectCompressed);
+
