+++ src/compressed-textures/ktx2/loadKTX2.ts
@@ -4,7 +4,7 @@ import { checkExtension } from '../../assets/utils/checkExtension';
 import { ExtensionType } from '../../extensions/Extensions';
 import { CompressedSource } from '../../rendering/renderers/shared/texture/sources/CompressedSource';
 import { getSupportedTextureFormats } from '../../rendering/renderers/shared/texture/utils/getSupportedTextureFormats';
-import { parseKTX } from './parseKTX';
+import { loadKTX2onWorker } from './worker/loadKTX2onWorker';
 
 import type { Loader } from '../../assets/loader/Loader';
 import type { LoaderParser } from '../../assets/loader/parsers/LoaderParser';
@@ -12,29 +12,25 @@ import type { ResolvedAsset } from '../../assets/types';
 import type { TextureSourceOptions } from '../../rendering/renderers/shared/texture/sources/TextureSource';
 import type { Texture } from '../../rendering/renderers/shared/texture/Texture';
 
-/** Loads KTX textures! */
-export const loadKTX = {
+/** Loads KTX2 textures! */
+export const loadKTX2 = {
     extension: {
         type: ExtensionType.LoadParser,
         priority: LoaderParserPriority.High,
     },
 
-    name: 'loadKTX',
+    name: 'loadKTX2',
 
     test(url: string): boolean
     {
-        return checkExtension(url, '.ktx');
+        return checkExtension(url, '.ktx2');
     },
 
     async load(url: string, _asset: ResolvedAsset, loader: Loader): Promise<Texture | Texture[]>
     {
         const supportedTextures = await getSupportedTextureFormats();
 
-        const ktxResponse = await fetch(url);
-
-        const ktxArrayBuffer = await ktxResponse.arrayBuffer();
-
-        const textureOptions = parseKTX(ktxArrayBuffer, supportedTextures);
+        const textureOptions = await loadKTX2onWorker(url, supportedTextures);
 
         const compressedTextureSource = new CompressedSource(textureOptions);
