+++ src/compressed-textures/ktx2/types.ts
