+++ src/compressed-textures/ktx2/utils/getTextureFormatFromKTXTexture.ts
