+++ src/compressed-textures/ktx2/worker/ktx.worker.ts
